+++ src/config_manager.hpp
@@ -24,11 +24,11 @@ struct app_setting {
 };
 
 struct bookmark {
-	int start;
-	int end;
+	long start;
+	long end;
 	wxString note;
 
-	bookmark(int s, int e, const wxString& n = wxEmptyString) : start{s}, end{e}, note{n} {
+	bookmark(long s, long e, const wxString& n = wxEmptyString) : start{s}, end{e}, note{n} {
 	}
 
 	bookmark() : start{0}, end{0}, note{wxEmptyString} {
@@ -103,22 +103,22 @@ public:
 	void remove_opened_document(const wxString& path);
 	wxArrayString get_opened_documents() const;
 	void clear_opened_documents();
-	void set_document_position(const wxString& path, int position);
-	int get_document_position(const wxString& path) const;
+	void set_document_position(const wxString& path, long position);
+	long get_document_position(const wxString& path) const;
 	void set_document_opened(const wxString& path, bool opened);
 	void remove_document_history(const wxString& path);
 	bool get_document_opened(const wxString& path) const;
 	wxArrayString get_all_opened_documents() const;
 	wxArrayString get_all_documents() const;
-	void add_bookmark(const wxString& path, int start, int end, const wxString& note = wxEmptyString);
-	void remove_bookmark(const wxString& path, int start, int end);
-	void toggle_bookmark(const wxString& path, int start, int end, const wxString& note = wxEmptyString);
-	void update_bookmark_note(const wxString& path, int start, int end, const wxString& note);
+	void add_bookmark(const wxString& path, long start, long end, const wxString& note = wxEmptyString);
+	void remove_bookmark(const wxString& path, long start, long end);
+	void toggle_bookmark(const wxString& path, long start, long end, const wxString& note = wxEmptyString);
+	void update_bookmark_note(const wxString& path, long start, long end, const wxString& note);
 	std::vector<bookmark> get_bookmarks(const wxString& path) const;
 	void clear_bookmarks(const wxString& path);
-	bookmark get_next_bookmark(const wxString& path, int current_position) const;
-	bookmark get_previous_bookmark(const wxString& path, int current_position) const;
-	bookmark get_closest_bookmark(const wxString& path, int current_position) const;
+	bookmark get_next_bookmark(const wxString& path, long current_position) const;
+	bookmark get_previous_bookmark(const wxString& path, long current_position) const;
+	bookmark get_closest_bookmark(const wxString& path, long current_position) const;
 	void set_document_format(const wxString& path, const wxString& format);
 	wxString get_document_format(const wxString& path) const;
 	bool needs_migration() const;
